BankAccounts and Cards Components (#155)
* Mergovane komponente za prikazivanje bankovnih racuna i kartica u jednu komponentu i popravljeno da se prikazuju sve kartice a ne samo jedna

* Obrisane stare BankAccount i Card komponente
